PackageManagerService -  Disable forward-locking.
Finsky has a habit of installing New and updated paid apps (and some others, though seemingly random) to the SDCard (native App2SD), regardless of the Install Location policy or the location of the app being updated. This has been identified as a conflict/incompatibility with forward-locking (the new "encryption" of Jellybean packages from Finksy) and (a) Non-ext4 internal storage and/or (b) devices with only one /storage mount (i.e. those with small userdata but true external SDCard).

This is solved by simply refusing forward-locking flags/requests on new installs. There are some caveats to consider:
(1) Packages that are already/previously installed to SDCard will be unassociated with the ASEC mount (DefaultContainerService) and cannot be moved. They must be uninstalled or (e.g.) Backed up with Titanium Backup then restored.
(2) Packages seem to prefer to be installed on Internal now, regardless of the install location policy. They have no problem being moved to/from SDCard once installed however.